+++ semcore/data-table/src/components/Body/Row.tsx
@@ -122,7 +122,9 @@ class RowRoot<UniqKeyType> extends Component<DataTableRowProps<UniqKeyType>, {},
                     checked={checked}
                     aria-labelledby={`${uid}_${ariaRowIndex}_1`}
                     onChange={this.handleSelectRow}
-                  />
+                  >
+                    <Checkbox.Value />
+                  </Checkbox>
                 </SCheckboxCell>,
               );
             }
@@ -180,7 +182,6 @@ class RowRoot<UniqKeyType> extends Component<DataTableRowProps<UniqKeyType>, {},
             interactive
             gridArea={accordionDataGridArea}
             duration={accordionDuration ?? 200}
-            zIndex={5}
           >
             <SCell
               aria-colindex={1}
@@ -192,8 +193,6 @@ class RowRoot<UniqKeyType> extends Component<DataTableRowProps<UniqKeyType>, {},
               columnIndex={1}
               // @ts-ignore
               column={{ name: ACCORDION }}
-              position='sticky'
-              left={0}
               w='100%'
               onKeyDown={this.handleBackFromAccordion}
             >