+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -1,5 +1,6 @@
 import { Box, ScreenReaderOnly, ScrollArea } from '@semcore/base-components';
 import { Component, createComponent, lastInteraction, Root, sstyled } from '@semcore/core';
+import canUseDOM from '@semcore/core/lib/utils/canUseDOM';
 import i18nEnhance from '@semcore/core/lib/utils/enhances/i18nEnhance';
 import findComponent from '@semcore/core/lib/utils/findComponent';
 import { hasParent } from '@semcore/core/lib/utils/hasParent';
@@ -42,10 +43,15 @@ export const ROW_INDEX = Symbol('ROW_INDEX');
 
 const SCROLL_BAR_HEIGHT = 12;
 
-type State = {
+type State<
+  Data extends DataTableData,
+  UniqKey extends keyof Data[number],
+  UniqKeyType extends Data[number][UniqKey],
+> = {
   scrollTop: number;
   scrollDirection: 'down' | 'up';
   selectAllMessage: string;
+  shadowVertical: BodyPropsInner<Data, UniqKeyType>['shadowVertical'];
 };
 
 class DataTableRoot<
@@ -76,6 +82,7 @@ class DataTableRoot<
   private columns: DTColumn[] = [];
   private treeColumns: DTColumn[] = [];
   private hasGroups = false;
+  private hasFixedColumn = false;
 
   private focusedCell: [RowIndex, ColIndex] = [-1, -1];
 
@@ -85,12 +92,14 @@ class DataTableRoot<
   private tableRef = React.createRef<HTMLDivElement>();
   private headerRef = React.createRef<HTMLDivElement>();
   private spinnerRef = React.createRef<HTMLDivElement>();
+  private containerResizeEndTimeoutId: ReturnType<typeof setTimeout> | null = null;
 
   private gridAreaGroupMap = new Map<number, string>();
 
   private columnsSplitter = '/';
-  private rows: Array<DTRow<UniqKeyType> | DTRow<UniqKeyType>[]> = [];
-  private flatRows: DTRow<UniqKeyType>[] = [];
+  private tmpData: Data;
+  private calculatedRows: Array<DTRow<UniqKeyType> | DTRow<UniqKeyType>[]>;
+  private flatRows: DTRow<UniqKeyType>[];
 
   private selectAllMessageTimer = 0;
 
@@ -107,14 +116,16 @@ class DataTableRoot<
       this.treeColumns = cols[1];
     }
 
-    this.rows = this.calculateRows();
-    this.flatRows = this.rows.flat();
+    this.calculatedRows = this.getRows();
+    this.flatRows = this.calculatedRows.flat();
+    this.tmpData = props.data;
   }
 
-  state: State = {
+  state: State<Data, UniqKey, UniqKeyType> = {
     scrollTop: 0,
     scrollDirection: 'down',
     selectAllMessage: '',
+    shadowVertical: '',
   };
 
   uncontrolledProps() {
@@ -124,11 +135,16 @@ class DataTableRoot<
   }
 
   componentDidMount() {
-    const { headerProps, loading, selectedRows } = this.asProps;
-    if ((headerProps?.sticky && !headerProps.h) || loading || this.columns.some((c) => c.fixed)) {
+    const { headerProps, loading } = this.asProps;
+    if ((headerProps?.sticky && !headerProps.h) || loading || this.hasFixedColumn) {
       requestAnimationFrame(() => {
         this.forceUpdate();
+        this.calculateVerticalShadow();
       });
+    }
+
+    if (headerProps?.sticky && canUseDOM() && this.scrollDirection === 'horizontal') {
+      document.addEventListener('scroll', this.handleDocumentScroll);
     }
   }
 
@@ -138,12 +154,13 @@ class DataTableRoot<
       const cols = this.calculateColumnsFromConfig();
       this.columns = cols[0];
       this.treeColumns = cols[1];
-    }
-    if (prevProps.data !== data || prevProps.columns !== columns) {
-      this.rows = this.calculateRows();
-      this.flatRows = this.rows.flat();
 
       this.forceUpdate();
+    }
+    if (prevProps.data !== data || prevProps.columns !== columns) {
+      if (this.hasFixedColumn) {
+        this.calculateVerticalShadow();
+      }
     }
     if (prevProps.selectedRows !== selectedRows && selectedRows !== undefined) {
       if (prevProps.selectedRows.length < data.length && selectedRows.length === data.length) {
@@ -154,9 +171,15 @@ class DataTableRoot<
     }
   }
 
+  componentWillUnmount() {
+    if (canUseDOM()) {
+      document.removeEventListener('scroll', this.handleDocumentScroll);
+    }
+  }
+
   get totalRows() {
     const { totalRows, expandedRows } = this.asProps;
-    const flatRows = this.rows.flat();
+    const flatRows = this.getFlatRows();
 
     const expandedRowsCount = Array.from(expandedRows ?? []).reduce((acc, rowKey) => {
       const dtRow = flatRows.find((el) => el[UNIQ_ROW_KEY] === rowKey);
@@ -177,7 +200,7 @@ class DataTableRoot<
       return totalRows + expandedRowsCount;
     }
 
-    const rows = this.rows.reduce((acc, item) => {
+    const rows = this.getRows().reduce((acc, item) => {
       acc = acc + 1;
 
       if (Array.isArray(item)) {
@@ -202,6 +225,28 @@ class DataTableRoot<
     };
   }
 
+  get scrollDirection() {
+    const { w, wMax, h, hMax } = this.asProps;
+
+    let scrollDirection: 'both' | 'horizontal' | 'vertical' | undefined = undefined;
+    const hasWidthSettings = Boolean(w) || Boolean(wMax);
+    const hasHeightSettings = (Boolean(h) && h !== 'fit-content') || Boolean(hMax);
+
+    if (hasWidthSettings && !hasHeightSettings) {
+      scrollDirection = 'horizontal';
+    } else if (hasHeightSettings && !hasWidthSettings) {
+      scrollDirection = 'vertical';
+    } else if (hasWidthSettings && hasHeightSettings) {
+      scrollDirection = 'both';
+    }
+
+    return scrollDirection;
+  }
+
+  get isDataEmpty() {
+    return this.asProps.data.length === 0;
+  }
+
   getHeadProps(): HeadPropsInner<Data, UniqKey, UniqKeyType> {
     const {
       use,
@@ -216,6 +261,7 @@ class DataTableRoot<
       sideIndents,
     } = this.asProps;
     const { gridTemplateColumns, gridTemplateAreas } = this.gridSettings;
+    const { shadowVertical } = this.state;
 
     return {
       ...headerProps,
@@ -235,9 +281,9 @@ class DataTableRoot<
       sideIndents,
       totalRows: this.totalRows,
       selectedRows,
-      flatRows: this.flatRows,
+      flatRows: this.getFlatRows(),
       onChangeSelectAll: (value, e) => {
-        const mappedFlatRows = this.flatRows.map((r) => r[UNIQ_ROW_KEY]);
+        const mappedFlatRows = this.getFlatRows().map((r) => r[UNIQ_ROW_KEY]);
         const selectedRowsSet = new Set(selectedRows);
 
         if (value) {
@@ -250,10 +296,13 @@ class DataTableRoot<
       },
       getFixedStyle: this.getFixedStyle,
       onCellClick: this.handleCellClick,
+      shadowVertical,
+      scrollDirection: this.scrollDirection,
+      isDataEmpty: this.isDataEmpty,
     };
   }
 
-  getBodyProps(): BodyPropsInner<UniqKeyType> {
+  getBodyProps(): BodyPropsInner<Data, UniqKeyType> {
     const {
       use,
       compact,
@@ -271,14 +320,16 @@ class DataTableRoot<
       accordionDuration,
       accordionMode,
       data: rawData,
+      renderCellOverlay,
     } = this.asProps;
     const { gridTemplateColumns, gridTemplateAreas } = this.gridSettings;
+    const { shadowVertical } = this.state;
     return {
       accordionDuration,
       accordionMode,
       columns: this.columns,
-      rows: this.rows,
-      flatRows: this.flatRows,
+      rows: this.getRows(),
+      flatRows: this.getFlatRows(),
       use,
       compact: Boolean(compact),
       gridTemplateColumns,
@@ -299,8 +350,8 @@ class DataTableRoot<
       virtualScroll,
       hasGroups: this.hasGroups,
       uid,
-      rowProps,
-      renderCell,
+      rowProps: this.getRows().length > 0 ? rowProps : undefined,
+      renderCell: this.getRows().length > 0 ? renderCell : undefined,
       renderEmptyData,
       sideIndents,
       selectedRows,
@@ -308,8 +359,47 @@ class DataTableRoot<
       getFixedStyle: this.getFixedStyle,
       onCellClick: this.handleCellClick,
       rawData,
+      shadowVertical,
+      renderCellOverlay,
     };
   }
+
+  handleDocumentScroll = trottle(() => {
+    const tableContainer = this.tableContainerRef.current;
+    if (!tableContainer) return;
+
+    const tableContainerTop = tableContainer.getBoundingClientRect().top;
+    const { headerProps } = this.asProps;
+    const headerContainer = this.headerRef.current;
+    const elements = headerContainer?.querySelectorAll('[role="columnheader"], [data-ui-name="Head.Group"]');
+    const top = tableContainerTop - (headerProps?.top ?? 0);
+    const headerScrollBar = headerProps?.withScrollBar
+      ? this.scrollAreaRef.current?.querySelector(`[role=scrollbar][aria-orientation=horizontal]`)
+      : undefined;
+
+    if (top && top < 0) {
+      const translate = `translateY(${Math.abs(top)}px)`;
+      elements?.forEach((column) => {
+        if (column instanceof HTMLElement) {
+          column.style.setProperty('transform', translate);
+        }
+      });
+
+      if (headerScrollBar instanceof HTMLElement) {
+        headerScrollBar.style.setProperty('transform', translate);
+      }
+    } else {
+      elements?.forEach((column) => {
+        if (column instanceof HTMLElement) {
+          column.style.removeProperty('transform');
+        }
+      });
+
+      if (headerScrollBar instanceof HTMLElement) {
+        headerScrollBar.style.removeProperty('transform');
+      }
+    }
+  });
 
   handleCellClick = (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow<UniqKeyType> }) => {
     if (lastInteraction.isMouse()) {
@@ -386,9 +476,7 @@ class DataTableRoot<
     if (expandedRows.has(expandedRow[UNIQ_ROW_KEY])) {
       expandedRows.delete(expandedRow[UNIQ_ROW_KEY]);
 
-      setTimeout(() => {
-        this.handlers.expandedRows(new Set([...expandedRows]));
-      }, 300);
+      this.handlers.expandedRows(new Set([...expandedRows]));
       onAccordionToggle?.('close', expandedRow[UNIQ_ROW_KEY], expandedRow[ROW_INDEX]);
     } else {
       expandedRows.add(expandedRow[UNIQ_ROW_KEY]);
@@ -564,15 +652,51 @@ class DataTableRoot<
   };
 
   handleScroll = trottle((e) => {
-    const scrollTop = e.target.scrollTop;
-    const scrollDirection = scrollTop > this.state.scrollTop ? 'down' : 'up';
-    this.setState({ scrollTop, scrollDirection });
+    if (this.asProps.virtualScroll) {
+      const scrollTop = e.target.scrollTop;
+      const scrollDirection = scrollTop > this.state.scrollTop ? 'down' : 'up';
+      this.setState({ scrollTop, scrollDirection });
+    }
+
+    if (this.hasFixedColumn) {
+      this.calculateVerticalShadow();
+    }
   });
+
+  calculateVerticalShadow = () => {
+    if (!this.tableContainerRef.current) return;
+
+    const { scrollWidth, clientWidth, scrollLeft } =
+            this.tableContainerRef.current;
+    const maxScrollRight = scrollWidth - clientWidth;
+
+    const roundedScroll = Math.round(scrollLeft);
+    const roundedMaxScroll = Math.round(maxScrollRight);
+    let shadow: BodyPropsInner<Data, UniqKeyType>['shadowVertical'] = '';
+    // not scroll
+    if (roundedMaxScroll <= 0) {
+      // start scroll
+    } else if (roundedScroll <= 0) {
+      shadow = 'end';
+      // end scroll
+    } else if (roundedScroll >= roundedMaxScroll) {
+      shadow = 'start';
+      // median scroll
+    } else if (roundedScroll > 0) {
+      shadow = 'median';
+    }
+
+    this.setState({ shadowVertical: shadow });
+  };
 
   handleFocus = (e: React.FocusEvent<HTMLElement, HTMLElement>) => {
     if (this.asProps.loading) {
       this.spinnerRef.current?.focus();
       e.currentTarget.setAttribute('tabIndex', '-1');
+
+      if (this.isDataEmpty) {
+        this.headerRef.current?.setAttribute('tabIndex', '-1');
+      }
     } else if (
       (!e.relatedTarget || !isFocusInside(e.currentTarget, e.relatedTarget)) &&
       lastInteraction.isKeyboard()
@@ -617,6 +741,10 @@ class DataTableRoot<
         }
       }
 
+      if (this.isDataEmpty) {
+        this.headerRef.current?.setAttribute('tabIndex', '-1');
+      }
+
       e.currentTarget.setAttribute('tabIndex', '-1');
     }
   };
@@ -633,6 +761,10 @@ class DataTableRoot<
     ) {
       this.setInert(false);
       tableElement.setAttribute('tabIndex', '0');
+
+      if (this.isDataEmpty) {
+        this.headerRef.current?.setAttribute('tabIndex', '0');
+      }
     }
   };
 
@@ -642,6 +774,14 @@ class DataTableRoot<
 
   handleBackFromAccordion = (cellIndex: number) => {
     this.changeFocusCell(-1, cellIndex === -1 ? 0 : cellIndex, 'up');
+  };
+
+  handleContainerResizeEnd = () => {
+    if (this.containerResizeEndTimeoutId) {
+      clearTimeout(this.containerResizeEndTimeoutId);
+    }
+
+    this.containerResizeEndTimeoutId = setTimeout(this.calculateVerticalShadow, 0);
   };
 
   render() {
@@ -683,18 +823,6 @@ class DataTableRoot<
       gridTemplateRows = `auto auto repeat(${this.totalRows}, minmax(${virtualScroll.rowHeight}px, auto)`;
     }
 
-    let scrollDirection: 'both' | 'horizontal' | 'vertical' | undefined = undefined;
-    const hasWidthSettings = Boolean(w) || Boolean(wMax);
-    const hasHeightSettings = (Boolean(h) && h !== 'fit-content') || Boolean(hMax);
-
-    if (hasWidthSettings && !hasHeightSettings) {
-      scrollDirection = 'horizontal';
-    } else if (hasHeightSettings && !hasWidthSettings) {
-      scrollDirection = 'vertical';
-    } else if (hasWidthSettings && hasHeightSettings) {
-      scrollDirection = 'both';
-    }
-
     return sstyled(styles)(
       <ScrollArea
         leftOffset={offsetLeftSum}
@@ -710,13 +838,14 @@ class DataTableRoot<
         ref={this.scrollAreaRef}
         container={this.tableContainerRef}
         styles={scrollStyles}
-        onScroll={virtualScroll ? this.handleScroll : undefined}
+        onScroll={this.handleScroll}
         disableAutofocusToContent={true}
+        onResize={this.handleContainerResizeEnd}
       >
         <ScrollArea.Container
           tabIndex={-1}
           // @ts-ignore
-          scrollDirection={scrollDirection}
+          scrollDirection={this.scrollDirection}
           // @ts-ignore
           loading={loading}
           headerHeight={`${headerHeight}px`}
@@ -732,6 +861,7 @@ class DataTableRoot<
             tabIndex={0}
             onFocus={this.handleFocus}
             onBlur={this.handleBlur}
+            isDataEmpty={this.isDataEmpty}
             aria-rowcount={this.totalRows}
             aria-colcount={this.columns.length}
             gridTemplateColumns={gridTemplateColumns.join(' ')}
@@ -880,7 +1010,7 @@ class DataTableRoot<
     if (selectedRows) {
       const column: DTColumn = {
         name: SELECT_ALL.toString(),
-        gtcWidth: '40px',
+        gtcWidth: 'min-content',
         alignItems: 'flex-start',
         children: '',
       };
@@ -999,7 +1129,7 @@ class DataTableRoot<
     if (selectedRows) {
       const column: DTColumn = {
         name: SELECT_ALL.toString(),
-        gtcWidth: '40px',
+        gtcWidth: 'min-content',
         alignItems: 'flex-start',
         children: '',
       };
@@ -1029,6 +1159,10 @@ class DataTableRoot<
         parent,
       } as DTColumn;
 
+      if (column.fixed) {
+        this.hasFixedColumn = true;
+      }
+
       return column;
     };
 
@@ -1043,6 +1177,19 @@ class DataTableRoot<
       return 'columns' in child;
     };
 
+    const setShowShadows = (col: DTColumn, i: number): void => {
+      let prevCol = treeColumns[i - 1];
+      if ('columns' in prevCol && prevCol.columns) {
+        prevCol = prevCol.columns[prevCol.columns.length - 1];
+      }
+
+      if (prevCol.fixed && !col.fixed) {
+        prevCol.showShadowVertical = true;
+      } else if (!prevCol.fixed && col.fixed) {
+        col.showShadowVertical = true;
+      }
+    };
+
     columns.forEach((child, i) => {
       if (childIsColumn(child)) {
         const col = makeColumn(child);
@@ -1055,6 +1202,10 @@ class DataTableRoot<
 
         calculatedColumns.push(col);
         treeColumns.push(col);
+
+        if (i > 0) {
+          setShowShadows(col, i);
+        }
       } else if (childIsGroup(child)) {
         const Group = makeColumn(child);
         const childCount = child.columns.length;
@@ -1082,6 +1233,10 @@ class DataTableRoot<
 
           calculatedColumns.push(col);
 
+          if (isFirst && i > 0) {
+            setShowShadows(col, i);
+          }
+
           Group.columns?.push(col);
         });
 
@@ -1095,10 +1250,28 @@ class DataTableRoot<
     return [calculatedColumns, treeColumns];
   }
 
-  private calculateRows(): Array<DTRow<UniqKeyType>[] | DTRow<UniqKeyType>> {
+  private getFlatRows(): DTRow<UniqKeyType>[] {
+    const { data } = this.props;
+
+    if (this.tmpData === data && this.flatRows) {
+      return this.flatRows;
+    }
+
+    this.flatRows = this.getRows().flat();
+
+    return this.flatRows;
+  }
+
+  private getRows(): Array<DTRow<UniqKeyType>[] | DTRow<UniqKeyType>> {
     const columns = this.columns;
     // @ts-ignore
     const { data, uid, uniqueRowKey } = this.props;
+
+    if (this.tmpData === data) {
+      return this.calculatedRows;
+    }
+
+    this.tmpData = data;
 
     const rows: Array<DTRow<UniqKeyType>[] | DTRow<UniqKeyType>> = [];
     const columnNames = columns.map((column: DTColumn) => column.name);
@@ -1207,6 +1380,7 @@ class DataTableRoot<
       }
     });
 
+    this.calculatedRows = rows;
     return rows;
   }