+++ semcore/feature-popover/src/index.ts
@@ -1,0 +1,8 @@
+import type { FeaturePopoverProps, FeaturePopoverPopperProps } from './FeaturePopover.type';
+
+export { default } from './FeaturePopover';
+
+export type {
+  FeaturePopoverProps,
+  FeaturePopoverPopperProps,
+};