+++ stories/components/data-table/tests/examples/scroll-tests/multiple-scrolls.tsx
@@ -20,6 +20,7 @@ const Demo = () => {
           { name: 'cpc', children: 'CPC' },
           { name: 'vol', children: 'Vol.' },
         ]}
+        defaultGridTemplateColumnWidth='min-content'
       />
 
     </>