+++ semcore/slider/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.0.7] - 2025-08-29
+
+### Changed
+
+- Type description for `SliderProps`.
+
+## [16.0.6] - 2025-08-07
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/flex-box` [16.0.5 ~> 16.0.6], `@semcore/base-components` [16.1.2 ~> 16.2.0]).
+
 ## [16.0.5] - 2025-07-07
 
 ### Fixed